+++ app/Exceptions/Handler.php
@@ -22,7 +22,7 @@ class Handler extends ExceptionHandler
      * @var array
      */
     protected $dontReport = [
-        //
+        NotInActivityInstanceException::class
     ];
 
     /**